[step2] refactor: deck refactoring, playerCards 삭제, 영향받는 코드 수정
