+++ src/main/kotlin/blackjack/Main.kt
@@ -1,7 +1,6 @@
 package blackjack
 
 import blackjack.card.Deck
-import blackjack.card.SimpleCardCreator
 import blackjack.dealer.Dealer
 import blackjack.player.PlayerCreator
 import blackjack.ui.input.BlackjackInputView
@@ -10,7 +9,7 @@ import blackjack.ui.output.ResultView
 fun main() {
     val playerNames = BlackjackInputView.createPlayerNames()
     val players = PlayerCreator.create(playerNames)
-    val deck = Deck(SimpleCardCreator.startCard())
+    val deck = Deck.init()
     val dealer = Dealer(deck)
 
     val blackJack = BlackjackGame(players, dealer)
