--- src/main/kotlin/blackjack/card/PlayCards.kt
@@ -1,11 +1,0 @@
-package blackjack.card
-
-class PlayCards(vararg value: Card) {
-
-    private var _cards: MutableList<Card> = value.toMutableList()
-    val size = _cards.size
-
-    fun drawCard(): Card {
-        return _cards.removeFirst()
-    }
-}
